feature: deal with ogmarkup, latex, markdown and others
The builder now distinguishes the documents files according to their
extension that is `.up` is an ogmarkup document (parse with ogmarkup),
`.md` a markdown file (pandoc) and `.tex` is a LaTex file (which is not
parse but included into the final document). A file with an unknown
extension in wrapped into a verbatim environment.